Move addUsdRate to prices.js
